Fix: Change baseline_smoothing and baseline_quantile to Int #217
The parameters were being set as float in one place

Issue: #217

+++ mzroll/settingsform.cpp
@@ -205,8 +205,8 @@ void SettingsForm::updateSettingFormGUI() {
     maxNaturalAbundanceErr->setValue(settings->value("maxNaturalAbundanceErr").toDouble());
     maxIsotopeScanDiff->setValue(settings->value("maxIsotopeScanDiff").toDouble());
     minIsotopicCorrelation->setValue(settings->value("minIsotopicCorrelation").toDouble());
-    baseline_smoothing->setValue(settings->value("baseline_smoothing").toDouble());
-    baseline_quantile->setValue(settings->value("baseline_quantile").toDouble());
+    baseline_smoothing->setValue(settings->value("baseline_smoothing").toInt());
+    baseline_quantile->setValue(settings->value("baseline_quantile").toInt());
 
     //Upload Multiprocessing
     checkBoxMultiprocessing->setCheckState( (Qt::CheckState) settings->value("uploadMultiprocessing").toInt() );
